removing old version of the run script.
